entry_from_search
