+++ Cargo.toml
@@ -32,7 +32,7 @@ windows-test-threading = []
 
 [dependencies]
 lazy_static = "1"
-keyring-search = "1"
+keyring-search = "1.1.1"
 
 [target.'cfg(target_os = "macos")'.dependencies]
 security-framework = { version = "2.6", optional = true }
