+++ src/lib.rs
@@ -390,9 +390,9 @@ impl Entry {
     /// The outer map's key corresponds to the ID of the result from 1 to the length of the map.
     /// The inner map contains the keys and values of the metadata of the result, i.e.
     /// target, service, user, last modified/date written, etc. In the case of keyutils, the Linux
-    /// Kernel keystore provides IDs for all credentials, the user must know the ID of the credential 
+    /// Kernel keystore provides IDs for all credentials, the user must know the ID of the credential
     /// to manipulate and pass this value to `from_search_results`. Since keyutils only returns one
-    /// result, this is the only valid parameter. 
+    /// result, this is the only valid parameter.
     /// # Example
     /// First result:
     /// ```rust
@@ -427,12 +427,8 @@ impl Entry {
             Some(credential) => credential.1,
             None => return Err(Error::NoEntry),
         };
-        // values[0] = target
-        // values[1] = service
-        // values[2] = user
-        let values: [&String; 3] = default::get_entry_values(credential)?;
 
-        Self::new_with_target(values[0], values[1], values[2])
+        default::entry_from_search(credential)
     }
 }
 
@@ -447,8 +443,6 @@ doc_comment::doctest!("../README.md", readme);
 // Since iOS doesn't use any of these generics, we allow dead code.
 #[allow(dead_code)]
 mod tests {
-    use std::collections::HashSet;
-
     use super::{credential::CredentialApi, Entry, Error, Result};
 
     /// Create a platform-specific credential given the constructor, service, and user
@@ -570,158 +564,5 @@ mod tests {
             &entry,
             "このきれいな花は桜です",
         );
-    }
-
-    #[test]
-    fn test_search_no_duplicates() {
-        let names = vec![
-            generate_random_string(),
-            generate_random_string(),
-            generate_random_string(),
-        ];
-
-        let mut entries = Vec::new();
-
-        for name in &names {
-            entries.push(
-                Entry::new_with_target(&name, "test-service", "test-user")
-                    .expect("Error constructing entry"),
-            )
-        }
-
-        for entry in &entries {
-            entry
-                .set_password("test-password")
-                .expect("Error setting password")
-        }
-
-        let search_users = keyring_search::List::list_credentials(
-            &Entry::search_users("test"),
-            keyring_search::Limit::All,
-        );
-        let users_set: HashSet<&str> = search_users.lines().collect();
-        let search_services = keyring_search::List::list_credentials(
-            &Entry::search_services("test"),
-            keyring_search::Limit::All,
-        );
-        let services_set: HashSet<&str> = search_services.lines().collect();
-        let search_default = keyring_search::List::list_credentials(
-            &Entry::search("test"),
-            keyring_search::Limit::All,
-        );
-        let search_set: HashSet<&str> = search_default.lines().collect();
-
-        assert_eq!(users_set, services_set);
-        assert_eq!(users_set, search_set);
-        assert_eq!(services_set, search_set);
-
-        for entry in &entries {
-            entry.delete_password().expect("error deleting entry")
-        }
-    }
-
-    #[test]
-    fn test_entry_from_search() {
-        let name = generate_random_string();
-        let password1 = "password1";
-        let password2 = "password2";
-        let entry = Entry::new_with_target(&name, "test-service", "test-user")
-            .expect("Error creating new entry");
-        entry
-            .set_password(password1)
-            .expect("error setting password1");
-
-        let old_password = entry
-            .get_password()
-            .expect("failed to get password from old entry");
-        let results = &Entry::search(&name);
-
-        let result_entry =
-            Entry::from_search_results(results, 1).expect("Failed to create entry from results");
-        result_entry
-            .set_password(password2)
-            .expect("error setting password2");
-
-        let new_password = result_entry
-            .get_password()
-            .expect("Failed to get password from new entry");
-
-        assert_eq!(password1, old_password);
-        assert_eq!(password2, new_password);
-
-        result_entry
-            .delete_password()
-            .expect("Failed to delete new entry");
-        let e = entry.delete_password().unwrap_err();
-
-        assert!(matches!(e, Error::NoEntry));
-    }
-
-    #[test]
-    fn test_entries_from_search() {
-        let names = [
-            generate_random_string(),
-            generate_random_string(),
-            generate_random_string(),
-        ];
-
-        let mut entries: Vec<Entry> = vec![];
-        let password1 = "password1";
-        let password2 = "password2";
-        for name in names {
-            let entry = Entry::new_with_target(&name, "test-service", "test-user")
-                .expect("Error creating new entry");
-            entry
-                .set_password(password1)
-                .expect("error setting password1");
-            entries.push(entry);
-        }
-
-        let mut old_passwords = vec![];
-
-        for entry in &entries {
-            let old_password = entry
-                .get_password()
-                .expect("failed to get password from old entry");
-            old_passwords.push(old_password);
-        }
-
-        let result = &Entry::search("test");
-
-        let size = result
-            .as_ref()
-            .expect("Error getting size of outer map")
-            .keys()
-            .len(); 
-        let mut result_entries: Vec<Entry> = vec![];
-        for index in 1..=size {
-            let msg = format!("Failed to create entry at index: {index}");
-            let entry = Entry::from_search_results(result, index).expect(&msg);
-            entry
-                .set_password(&password2)
-                .expect("Error setting new password");
-            result_entries.push(entry);
-        }
-
-        let mut new_passwords = vec![];
-
-        for entry in &result_entries {
-            let new_password = entry.get_password().expect("error getting new password");
-            new_passwords.push(new_password);
-        }
-
-        for i in 0..new_passwords.len() {
-            assert_eq!(password1, old_passwords[i]);
-            assert_eq!(password2, new_passwords[i]);
-        }
-
-        for entry in &result_entries {
-            entry.delete_password().expect("Error deleting password");
-        }
-
-        for entry in &entries {
-            let e = entry.delete_password().unwrap_err();
-            assert!(matches!(e, Error::NoEntry));
-        }
     }
 }
