+++ src/windows.rs
@@ -43,6 +43,8 @@ use windows_sys::Win32::Security::Credentials::{
     CRED_MAX_USERNAME_LENGTH, CRED_PERSIST_ENTERPRISE, CRED_TYPE_GENERIC,
 };
 
+use crate::Entry;
+
 use super::credential::{Credential, CredentialApi, CredentialBuilder, CredentialBuilderApi};
 use super::error::{Error as ErrorCode, Result};
 
@@ -365,9 +367,7 @@ unsafe fn from_wstr(ws: *const u16) -> String {
     String::from_utf16_lossy(slice)
 }
 
-pub fn get_entry_values(
-    credential: &std::collections::HashMap<String, String>,
-) -> Result<[&String; 3]> {
+pub fn entry_from_search(credential: &std::collections::HashMap<String, String>) -> Result<Entry> {
     let target = if let Some(target) = credential.get(&"Target".to_string()) {
         target
     } else {
@@ -392,7 +392,15 @@ pub fn get_entry_values(
             "No user key found in credential".to_string(),
         ));
     };
-    Ok([target, comment, user])
+
+    let wincredential = Box::new(WinCredential {
+        username: user.to_string(),
+        target_name: target.to_string(),
+        target_alias: "".to_string(),
+        comment: comment.to_string(),
+    });
+
+    Ok(Entry::new_with_credential(wincredential))
 }
 
 /// Windows error codes are `DWORDS` which are 32-bit unsigned ints.
@@ -435,6 +443,8 @@ fn wrap(code: u32) -> Box<dyn std::error::Error + Send + Sync> {
 
 #[cfg(test)]
 mod tests {
+    use std::collections::HashSet;
+
     use super::*;
 
     use crate::credential::CredentialPersistence;
@@ -625,5 +635,157 @@ mod tests {
             .delete_password()
             .expect("Couldn't delete get-credential");
         assert!(matches!(entry.get_password(), Err(ErrorCode::NoEntry)));
+    }
+    #[test]
+    fn test_search_no_duplicates() {
+        let names = vec![
+            generate_random_string(),
+            generate_random_string(),
+            generate_random_string(),
+        ];
+
+        let mut entries = Vec::new();
+
+        for name in &names {
+            entries.push(
+                Entry::new_with_target(&name, "test-service", "test-user")
+                    .expect("Error constructing entry"),
+            )
+        }
+
+        for entry in &entries {
+            entry
+                .set_password("test-password")
+                .expect("Error setting password")
+        }
+
+        let search_users = keyring_search::List::list_credentials(
+            &Entry::search_users("test"),
+            keyring_search::Limit::All,
+        );
+        let users_set: HashSet<&str> = search_users.lines().collect();
+        let search_services = keyring_search::List::list_credentials(
+            &Entry::search_services("test"),
+            keyring_search::Limit::All,
+        );
+        let services_set: HashSet<&str> = search_services.lines().collect();
+        let search_default = keyring_search::List::list_credentials(
+            &Entry::search("test"),
+            keyring_search::Limit::All,
+        );
+        let search_set: HashSet<&str> = search_default.lines().collect();
+
+        assert_eq!(users_set, services_set);
+        assert_eq!(users_set, search_set);
+        assert_eq!(services_set, search_set);
+
+        for entry in &entries {
+            entry.delete_password().expect("error deleting entry")
+        }
+    }
+
+    #[test]
+    fn test_entry_from_search() {
+        let name = generate_random_string();
+        let password1 = "password1";
+        let password2 = "password2";
+        let entry = Entry::new_with_target(&name, "test-service", "test-user")
+            .expect("Error creating new entry");
+        entry
+            .set_password(password1)
+            .expect("error setting password1");
+
+        let old_password = entry
+            .get_password()
+            .expect("failed to get password from old entry");
+        let results = &Entry::search(&name);
+
+        let result_entry =
+            Entry::from_search_results(results, 1).expect("Failed to create entry from results");
+        result_entry
+            .set_password(password2)
+            .expect("error setting password2");
+
+        let new_password = result_entry
+            .get_password()
+            .expect("Failed to get password from new entry");
+
+        assert_eq!(password1, old_password);
+        assert_eq!(password2, new_password);
+
+        result_entry
+            .delete_password()
+            .expect("Failed to delete new entry");
+        let e = entry.delete_password().unwrap_err();
+
+        assert!(matches!(e, ErrorCode::NoEntry));
+    }
+
+    #[test]
+    fn test_entries_from_search() {
+        let names = [
+            generate_random_string(),
+            generate_random_string(),
+            generate_random_string(),
+        ];
+
+        let mut entries: Vec<Entry> = vec![];
+        let password1 = "password1";
+        let password2 = "password2";
+        for name in names {
+            let entry = Entry::new_with_target(&name, "test-service", "test-user")
+                .expect("Error creating new entry");
+            entry
+                .set_password(password1)
+                .expect("error setting password1");
+            entries.push(entry);
+        }
+
+        let mut old_passwords = vec![];
+
+        for entry in &entries {
+            let old_password = entry
+                .get_password()
+                .expect("failed to get password from old entry");
+            old_passwords.push(old_password);
+        }
+
+        let result = &Entry::search("test");
+
+        let size = result
+            .as_ref()
+            .expect("Error getting size of outer map")
+            .keys()
+            .len();
+        let mut result_entries: Vec<Entry> = vec![];
+        for index in 1..=size {
+            let msg = format!("Failed to create entry at index: {index}");
+            let entry = Entry::from_search_results(result, index).expect(&msg);
+            entry
+                .set_password(&password2)
+                .expect("Error setting new password");
+            result_entries.push(entry);
+        }
+
+        let mut new_passwords = vec![];
+
+        for entry in &result_entries {
+            let new_password = entry.get_password().expect("error getting new password");
+            new_passwords.push(new_password);
+        }
+
+        for i in 0..new_passwords.len() {
+            assert_eq!(password1, old_passwords[i]);
+            assert_eq!(password2, new_passwords[i]);
+        }
+
+        for entry in &result_entries {
+            entry.delete_password().expect("Error deleting password");
+        }
+
+        for entry in &entries {
+            let e = entry.delete_password().unwrap_err();
+            assert!(matches!(e, ErrorCode::NoEntry));
+        }
     }
 }
